# docs/swagger.yaml
definitions:
...
  dto.ApiComment:
    properties:
      author:
        description: コメント投稿者ID（匿名の場合はnil）
        type: string
      color:
        description: コメント色（16進数カラーコード）
        type: string
      size:
        description: コメントサイズ
        enum:
        - big
        - medium
        - small
        type: string
      text:
        description: コメント本文
        type: string
      time:
        description: コメント表示時刻（秒）
        minimum: 0
        type: number
      type:
        description: コメント位置
        enum:
        - right
        - top
        - bottom
        type: string
    required:
    - color
    - size
    - text
    - time
    - type
    type: object
  dto.Capture:
    properties:
      _:
        type: string
      comment_delay:
        type: number
      created_at:
        type: string
      filename:
        type: string
      id:
        type: integer
      playback_position:
        type: number
      video_id:
        type: integer
    type: object
  dto.CaptureListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.Capture'
        type: array
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
  dto.CaptureSeriesInfo:
    properties:
      first_month:
        type: integer
      first_year:
        type: integer
      id:
        type: integer
      series_name_file:
        type: string
      syobocal_title_id:
        type: integer
      syobocal_title_name:
        type: string
      syobocal_title_name_en:
        type: string
    type: object
  dto.CapturesSeriesListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.CaptureSeriesInfo'
        type: array
    type: object
...
  dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
    type: object
...
  dto.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.SeriesListResponse:
    properties:
      series:
        items:
          $ref: '#/definitions/dto.SeriesResponse'
        type: array
      total:
        type: integer
    type: object
  dto.SeriesResponse:
    properties:
      comment: {}
      created_at:
        type: string
      first_end_month:
        type: integer
      first_end_year:
        type: integer
      first_month:
        type: integer
      first_year:
        type: integer
      id:
        type: integer
      series_name_file:
        type: string
      subtitles: {}
      syobocal_title_id:
        type: integer
      syobocal_title_name:
        type: string
      syobocal_title_name_en:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.SeriesWithVideosResponse:
    properties:
      series:
        $ref: '#/definitions/dto.SeriesResponse'
      videos:
        items:
          $ref: '#/definitions/dto.VideoResponse'
        type: array
    type: object
  dto.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.VideoResponse:
    properties:
      channel_id:
        description: Syobocal チャンネルID（ChID）
        type: integer
      comments:
        items:
          $ref: '#/definitions/dto.ApiComment'
        type: array
      created_at:
        type: string
      description:
        type: string
      duration:
        type: number
      episode:
        description: エピソード番号
        type: integer
      file_name:
        type: string
      file_size:
        type: integer
      folder_id:
        description: Folderテーブルへの外部キー
        type: integer
      id:
        type: integer
      is_deleted:
        type: boolean
      jikkyo_comment_count:
        type: integer
      jikkyo_date:
        type: string
      liked:
        type: boolean
      prog_end_time:
        description: 放送終了時刻
        type: string
      prog_start_time:
        description: 放送開始時刻
        type: string
      screenshot_file_path:
        type: string
      series:
        allOf:
        - $ref: '#/definitions/entity.Series'
        description: Series関連付け
      series_id:
        description: Seriesテーブルへの外部キー
        type: integer
      src:
        type: string
      status:
        description: ready, processing, error
        type: string
      subtitle:
        description: エピソードサブタイトル
        type: string
      thumbnail_info:
        $ref: '#/definitions/entity.ThumbnailInfo'
      updated_at:
        type: string
      views:
        type: integer
    type: object
...
  entity.Capture:
    properties:
      _:
        type: string
      comment_delay:
        type: number
      created_at:
        type: string
      filename:
        type: string
      id:
        type: integer
      playback_position:
        type: number
      video_id:
        type: integer
    type: object
  entity.JSONMap:
    additionalProperties: true
    type: object
  entity.Series:
    properties:
      comment:
        $ref: '#/definitions/entity.JSONMap'
      created_at:
        type: string
      first_end_month:
        type: integer
      first_end_year:
        type: integer
      first_month:
        type: integer
      first_year:
        type: integer
      id:
        type: integer
      series_name_file:
        description: '{title} 部分'
        type: string
      subtitles:
        items:
          additionalProperties: true
          type: object
        type: array
      syobocal_title_id:
        type: integer
      syobocal_title_name:
        type: string
      syobocal_title_name_en:
        type: string
      updated_at:
        type: string
    type: object
  entity.ThumbnailInfo:
    properties:
      generated_at:
        type: string
      height:
        type: integer
      width:
        type: integer
    type: object
...
  gin.H:
    additionalProperties: {}
    type: object
info:
  contact: {}
  description: Watch commentPlayer API documentation
  title: commentPlayer API
  version: 1.0.0
paths:
...
  /api/v1/captures:
    get:
      description: キャプチャ一覧をページネーション付きで取得します。series_id でシリーズ別フィルター、sort_key と sort_order
        でソートをカスタマイズできます
      parameters:
      - description: ビデオID（フィルタリング用）
        in: query
        name: video_id
        type: integer
      - description: シリーズID（フィルタリング用）
        in: query
        name: series_id
        type: integer
      - default: 1
        description: ページ番号
        in: query
        name: page
        type: integer
      - default: 20
        description: 1ページあたりのアイテム数
        in: query
        name: limit
        type: integer
      - default: created_at
        description: ソート対象フィールド (id または created_at)
        enum:
        - id
        - created_at
        in: query
        name: sort_key
        type: string
      - default: desc
        description: ソート順序 (asc または desc)
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CaptureListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: キャプチャ一覧を取得
      tags:
      - Captures
    post:
      description: 新しいキャプチャを作成します
      parameters:
      - description: キャプチャファイル
        in: formData
        name: file
        required: true
        type: file
      - description: ビデオID
        in: formData
        name: video_id
        required: true
        type: integer
      - description: ビデオの再生位置（秒）
        in: formData
        name: playback_position
        type: number
      - description: コメント遅延秒数
        in: formData
        name: comment_delay
        type: number
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Capture'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: キャプチャを作成
      tags:
      - Captures
  /api/v1/captures/{id}:
    delete:
      description: キャプチャをファイルシステムとDBから削除します
      parameters:
      - description: キャプチャID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: キャプチャを削除
      tags:
      - Captures
    get:
      description: キャプチャをIDで取得します
      parameters:
      - description: キャプチャID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Capture'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: キャプチャを取得
      tags:
      - Captures
  /api/v1/captures/series:
    get:
      description: キャプチャが存在するシリーズのみの一覧を返します
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CapturesSeriesListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: キャプチャ対応シリーズ一覧を取得
      tags:
      - Captures
...
  /api/v1/series:
    get:
      description: 全てのシリーズの一覧を取得します
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SeriesListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: シリーズ一覧を取得
      tags:
      - Series
  /api/v1/series/{id}:
    get:
      description: 指定されたシリーズIDに関連する全てのビデオを取得します
      parameters:
      - description: シリーズID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SeriesWithVideosResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: シリーズとそのビデオを取得
      tags:
      - Series
...
schemes:
- http
swagger: "2.0"